+++ DenisIlimbaev/HomeTask11And12/CosmeticCosm.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace HomeTask11
+{
+    enum CategoryCosm
+    {
+        Man,
+        Woman
+    }
+}
