+++ DenisIlimbaev/HomeTask11And12/Package.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace HomeTask11
+{
+    enum Package
+    {
+        Carton,
+        Paper,
+        Bottle,
+        Metal
+    }
+}
